requirements.txt
selenium~=4.31.0
undetected-chromedriver~=3.5.5
python-dotenv~=1.1.0
pydantic~=2.11.3